+++ client/verta/verta/_internal_utils/_pip_requirements_utils.py
@@ -11,7 +11,7 @@ import warnings
 
 import cloudpickle
 
-from ..external import six
+from .._vendored import six
 
 from .. import __about__
