+++ client/verta/verta/data_types/_float_histogram.py
@@ -2,7 +2,7 @@
 
 from __future__ import division
 
-from ..external import six
+from .._vendored import six
 
 from .._internal_utils import arg_handler
 from .._internal_utils.importer import maybe_dependency
