+++ client/verta/verta/data_types/_numeric_value.py
@@ -4,7 +4,7 @@ from __future__ import division
 
 import numbers
 
-from ..external import six
+from .._vendored import six
 
 from .._internal_utils import arg_handler
