+++ client/verta/verta/data_types/_string_value.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from ..external import six
+from .._vendored import six
 
 from .._internal_utils import arg_handler
