+++ client/verta/verta/dataset/_s3.py
@@ -7,7 +7,7 @@ import pathlib
 import tempfile
 from urllib.parse import urlparse
 
-from ..external import six
+from .._vendored import six
 
 from .._protos.public.modeldb.versioning import (
     VersioningService_pb2 as _VersioningService,
