+++ client/verta/verta/endpoint/autoscaling/metrics.py
@@ -3,7 +3,7 @@
 
 import abc
 
-from ...external import six
+from ..._vendored import six
 
 
 @six.add_metaclass(abc.ABCMeta)
