+++ client/verta/verta/endpoint/update/rules.py
@@ -4,7 +4,7 @@
 import abc
 import json
 
-from ...external import six
+from ..._vendored import six
 
 
 @six.add_metaclass(abc.ABCMeta)
