+++ client/verta/verta/environment/_python.py
@@ -5,7 +5,7 @@ import logging
 import os
 import sys
 
-from ..external import six
+from .._vendored import six
 
 from .._protos.public.modeldb.versioning import (
     VersioningService_pb2 as _VersioningService,
