+++ client/verta/verta/tracking/entities/_experimentrun.py
@@ -24,7 +24,7 @@ from verta._protos.public.modeldb import (
     ExperimentRunService_pb2 as _ExperimentRunService,
 )
 
-from verta.external import six
+from verta._vendored import six
 
 from verta._internal_utils import (
     _artifact_utils,
