+++ client/verta/verta/tracking/entities/_experimentruns.py
@@ -9,7 +9,7 @@ from verta._protos.public.modeldb import (
     ExperimentRunService_pb2 as _ExperimentRunService,
 )
 
-from verta.external import six
+from verta._vendored import six
 
 from verta._bases import _LazyList
 from verta._internal_utils import importer
